Splits Root and App components
Also, adds work type route.
