+++ ui/app/components/AppComponent/AppComponent.css
@@ -1,7 +1,3 @@
-.app {
-  margin-top: 70px;
-}
-
 .label {
   color: #666;
   padding: 6px;
