+++ ui/app/components/AppComponent/index.js
@@ -1,0 +1,3 @@
+import AppComponent from './AppComponent';
+
+export default AppComponent;
