+++ ui/app/components/CPUSamplingComponent/CPUSamplingComponent.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+export class CPUSamplingComponent extends Component {
+  render () {
+    return (
+      <div>
+        <h1>I am CPU Sampling Component</h1>
+      </div>
+    );
+  }
+}
+
+export default CPUSamplingComponent;
