+++ ui/app/components/CPUSamplingComponent/index.js
@@ -1,0 +1,3 @@
+import CPUSamplingComponent from './CPUSamplingComponent';
+
+export default CPUSamplingComponent;
