+++ ui/app/components/ProfileListComponent/ProfileListComponent.css
@@ -54,7 +54,6 @@
   }
 }
 
-.worktype {
-  display: inline-block;
+.workType {
   margin-right: 3px;
 }
